requirements.txt
argparse
langchain
lxml
pymongo
openai
fastapi
uvicorn
pymongo
pydantic
langchain-openai
langchain-community
langchain-text-splitters
requests
tiktoken
langchain-mongodb
jq
python-dotenv